requirements.txt
django==4.2.2
gunicorn==20.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.4
